requirements.txt
django==5.0.4 
djangorestframework==3.15.1 
drf-yasg==1.21.7
setuptools==69.2.0
djangorestframework-simplejwt==5.3.1